TASK_2/combine_front_back_mqtt/requirements.txt
Flask==2.3.3
paho-mqtt==1.6.1
Flask-SocketIO==5.5.1